+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This is a changelog for Piwik platform developers. All changes for our HTTP API'
 ### Breaking Changes
 * The deprecated method `Period::factory()` has been removed. Use `Period\Factory` instead.
 * The deprecated method `Config::getConfigSuperUserForBackwardCompatibility()` has been removed.
+* The deprecated methods `MenuAdmin::addEntry()` and `MenuAdmin::removeEntry()` have been removed. Use `Piwik\Plugin\Menu` instead.
 
 ## Piwik 2.11.0